refactor: 서비스를 service 패키지로 이동 (#108)
