+++ src/main/java/com/octagram/pollet/notification/application/service/NotificationService.java
@@ -1,4 +1,4 @@
-package com.octagram.pollet.notification.application;
+package com.octagram.pollet.notification.application.service;
 
 import com.octagram.pollet.notification.domain.repository.NotificationRepository;
 import lombok.RequiredArgsConstructor;
